% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/couleurs_hug.R
\name{couleurs_hug}
\alias{couleurs_hug}
\title{Couleurs officielles HUG}
\usage{
couleurs_hug(type = "principales")
}
\arguments{
\item{type}{character. "principales", "secondaires bleu" ou "secondaires rouge"}
}
\description{
Couleurs officielles HUG
}
\examples{
couleurs_hug("principales")
}
