% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct_dates.R
\name{distinct_dates}
\alias{distinct_dates}
\title{Reduit les lignes de dates}
\usage{
distinct_dates(
  data,
  groups,
  variable_date_debut,
  variable_date_fin,
  tronquer_debut = NULL,
  tronquer_fin = NULL,
  tolerance_weekend = TRUE,
  tolerance_nb_jours = 1
)
}
\arguments{
\item{data}{une table contenant au moins une colonne date de debut et une colonne date de fin (voir autres arugments)}

\item{groups}{le(s) (differents) groupe(s) dont il faut tenir compte variable d'identification (ex: Matricule, Statut)}

\item{variable_date_debut}{une seule variable qui indique la date de debut (debut agent, debut absences, ...)}

\item{variable_date_fin}{une seule variable qui indique la date de fin (fin agent, fin absences, ...)}

\item{tronquer_debut}{Date entre guillemets `ymd("2018-01-01")` dans le format "YYYY-MM-JJ" pour tronquer les dates de debut (utile pour le calcul du nombre de jours)}

\item{tronquer_fin}{Date entre guillemets `ymd("2018-01-01")` dans le format "YYYY-MM-JJ" pour tronquer les dates de fin (utile pour le calcul du nombre de jours)}

\item{tolerance_weekend}{si TRUE, alors tolere une difference de jours deux lignes si cette difference est uniquement sur le weekend (date de fin un vendredi/samedi/dimanche et date de debut suivante est un lundi)}

\item{tolerance_nb_jours}{default 1. Nombre de jours de tolerance entre la date de fin et la prochaine date de debut (par exemple 31 jours pour les contrats agents ou 1 jour pour les absences a la suite)}
}
\value{
une table reduite
}
\description{
Reduire les lignes d'une table selon la date de debut et la date de fin. Utiles pour analyser les differentes saisies d'absences ou les differents contrats agents qui se suivent, etc.
}
\examples{
distinct_dates(data, MATRICULE, variable_date_debut = `date debut agent`, variable_date_fin = `date fin agent`, tolerance_nb_jours = 31)
# Multiple variables de groupes
distinct_date(data, vars(MATRICULE, STATUT), date_debut, date_fin )

}
