% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate_rows.R
\name{duplicate_rows}
\alias{duplicate_rows}
\title{Dupliquer le nombre de lignes}
\usage{
duplicate_rows(df, from, to, .new_name = ".date")
}
\arguments{
\item{df}{a data.frame}

\item{from}{variable date}

\item{to}{variable date}

\item{.data}{a strin}
}
\description{
Utile pour avoir
}
\examples{
X <- data.frame(id = 1:5,
  from = Sys.Date() + 1:5,
  to = Sys.Date() + c(1,3,5,8,10))
duplicate_rows(X, from, to)
}
