% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outilsSIRH-package.R
\docType{package}
\name{outilsSIRH-package}
\alias{outilsSIRH}
\alias{outilsSIRH-package}
\title{outilsSIRH: Analyser les donnees SIRH}
\description{
Ce package contient differentes fonctions utiles pour l'analyse de donnees du SIRH. 
    Ce package n'est absolument pas officiel et il est surtout une aide pour son auteur.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sirh-hug/outilsSIRH}
  \item Report bugs at \url{https://github.com/sirh-hug/outilsSIRH/issues}
}

}
\keyword{internal}
