% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_table.R
\name{print_table}
\alias{print_table}
\title{Formater un tableau croise}
\usage{
print_table(data, bg.col = "#a0d0ab", title = NULL, caption = "")
}
\arguments{
\item{bg.col}{color. Background color}

\item{title}{character.}

\item{caption}{character.}

\item{table}{a table.}
}
\description{
Formater un tableau croise
}
