% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hug.R
\name{theme_hug}
\alias{theme_hug}
\title{Theme HUG pour ggplot2}
\usage{
theme_hug()
}
\description{
Theme HUG pour ggplot2
}
\examples{
ggplot(mtcars) +
  aes(x = mpg, y = qsec) +
  geom_point() +
  theme_hug()
}
